third_party/blink/web_tests/platform/mac-mac11-arm64/external/wpt/resource-timing/iframe-sequence-of-events-expected.txt

This is a testharness.js-based test.
PASS An iframe should report its RT entry when the response is done and before it is completely loaded
FAIL An iframe should report separate RT entries if src changed assert_equals: expected 2 but got 1
Harness: the test ran to completion.
